Improve formatting of README.md
The features section now only says the names of the features that are in the extension.
Rather than displaying the description of the snippets in the features section, it is now shown in the snippets section.
I have created a section for the database browser that somebody with more knowledge than me could fill in.