DA-118: Modified: Retrieve existing document content when creating ne… (#92)
Current Behaviour: While we are creating a new document with the same name which already exist we are getting the new document with empty content and the current document’s content get erased. 

Expected Behaviour: While we are creating a new document with the same name which already exist we should get the document with existing content. 

- Closes #91 

Co-authored-by: Aayush Tyagi <aayushtyagi@Aayushs-MacBook-Pro.local>