Update to use Couchbase Node SDK v4 (#98)
## Describe the problem this PR is solving
Updates the extension to use Couchbase Node SDK v4. v4 of the SDK
utilises a different build system (cmake-js) and comes bundled with
binding files for it's native C++ code. These binding files are for Node
and we need to rebuild them for use with Electron.

## Short description of the changes
- Update Couchbase Node SDK to v4.2.0
- Add npm scripts to manage building these new binding files
- Remove old script to manage the old binding file process
- Add Developing.md to describe the bindings files, how to build them
and an OpenSSL/BoringSSL issue