Da#107 notify user when document has changed (#115)
## Describe the problem this PR is solving
After you open a document and leave it open, the content is never
updated. So it might not reflect the real content of the document
anymore.

## Short description of the changes
Whenever the document is on focus again, send a request to the cluster
and check if the CAS still is the same. If it is not, show a warning to
the user saying that the document has been updated and give him two
options "Keep Local Version", "Update to Latest Version" (Feel free to
give a better name to these buttons).

If the user saves a document, we will need to make the same
verification, but in this case, we will need to show an ERROR message
like: "There is a conflict while trying to save this document, as it was
also changed in the server. Would you like to load the server version or
overwrite the remote version with your changes? " and the two options:
"Load new remove version" and "Overwrite it with my changes"
<img width="1067" alt="Screenshot 2023-02-10 at 5 30 00 PM"
src="https://user-images.githubusercontent.com/42893909/218089850-b1232eb2-2829-47d5-936f-f2e9df197545.png">
<img width="1078" alt="Screenshot 2023-02-10 at 5 30 57 PM"
src="https://user-images.githubusercontent.com/42893909/218089882-4162b5ef-83c8-4497-95da-5fc5610e5d59.png">